requirements.txt
streamlit
mistralai
qdrant-client
python-dotenv